# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API para sitio web corporativo de jardinería B2B
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type QuoteInputClientType = typeof QuoteInputClientType[keyof typeof QuoteInputClientType];


export const QuoteInputClientType = {
  conjunto_residencial: 'conjunto_residencial',
  colegio: 'colegio',
  edificio: 'edificio',
  centro_comercial: 'centro_comercial',
  otro: 'otro',
} as const;

export type QuoteInputService = typeof QuoteInputService[keyof typeof QuoteInputService];


export const QuoteInputService = {
  mantenimiento: 'mantenimiento',
  diseno: 'diseno',
  instalacion: 'instalacion',
  poda: 'poda',
  fumigacion: 'fumigacion',
  otro: 'otro',
} as const;

export interface QuoteInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 2 */
  company: string;
  clientType: QuoteInputClientType;
  /** @minLength 7 */
  phone: string;
  /** @nullable */
  email?: string | null;
  service: QuoteInputService;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  area?: string | null;
  /** @minLength 10 */
  message: string;
}

export interface Quote {
  id: number;
  name: string;
  company: string;
  clientType: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  service: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  area?: string | null;
  message: string;
  createdAt: string;
}

export interface Project {
  id: number;
  title: string;
  category: string;
  location: string;
  description: string;
  imageUrl: string;
  year: number;
  /** @nullable */
  areaSqm?: number | null;
  tags?: string[];
}

export type PortfolioStatsByCategory = {[key: string]: number};

export interface PortfolioStats {
  totalProjects: number;
  totalClients: number;
  yearsExperience: number;
  totalAreaSqm: number;
  byCategory?: PortfolioStatsByCategory;
}

export type ListProjectsParams = {
category?: ListProjectsCategory;
};

export type ListProjectsCategory = typeof ListProjectsCategory[keyof typeof ListProjectsCategory];


export const ListProjectsCategory = {
  conjuntos_residenciales: 'conjuntos_residenciales',
  colegios: 'colegios',
  edificios: 'edificios',
  centros_comerciales: 'centros_comerciales',
} as const;