# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API para sitio web corporativo de jardinería B2B
 * OpenAPI spec version: 0.1.0
 */

export interface Project {
  id: number;
  title: string;
  category: string;
  location: string;
  description: string;
  imageUrl: string;
  year: number;
  /** @nullable */
  areaSqm?: number | null;
  tags?: string[];
}