# lib/api-zod/src/generated/types/quote.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API para sitio web corporativo de jardinería B2B
 * OpenAPI spec version: 0.1.0
 */

export interface Quote {
  id: number;
  name: string;
  company: string;
  clientType: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  service: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  area?: string | null;
  message: string;
  createdAt: string;
}