# lib/api-zod/src/generated/types/quoteInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API para sitio web corporativo de jardinería B2B
 * OpenAPI spec version: 0.1.0
 */
import type { QuoteInputClientType } from './quoteInputClientType';
import type { QuoteInputService } from './quoteInputService';

export interface QuoteInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 2 */
  company: string;
  clientType: QuoteInputClientType;
  /** @minLength 7 */
  phone: string;
  /** @nullable */
  email?: string | null;
  service: QuoteInputService;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  area?: string | null;
  /** @minLength 10 */
  message: string;
}